Update conventions after linting
Before publishing to MELPA, I had to complete a submission checklist. That
required me to run `package-lint`, which forced me to change from the Scheme and
Clojure style of using forward-slashes for namespaces to using dashes, which is
a bit upsetting.

But since this is the first package that I'm publishing, I'm moving forward with
the changes.